Handle repeated NETFILTER_CFG records
It is not clear what those records mean, but it seems we can simply put
them into an array for now.

Perhaps later we can have two-level map, with first level having tables
as keys and the second level having address families as keys and number
of entries as values. Although that would be against the spirit of
keeping the schema as close to the original log as possible.

Fixes #62